refactor: rename utils to getTilesRectFromCenter
